auth-server/requirements.txt
fastapi==0.115.12
uvicorn==0.34.0
python-jose==3.3.0
pydantic==2.10.6
